requirements.txt
Brotli==1.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
idna==3.7
inflate64==1.0.0
multivolumefile==0.2.3
psutil==5.9.8
py7zr==0.21.0
pybcj==1.0.2
pycryptodomex==3.20.0
pyppmd==1.1.0
python-magic==0.4.27
pyzstd==0.16.0
requests==2.32.3
texttable==1.7.0
urllib3==2.2.1
